+++ src/backend/core/recording/event/__init__.py
@@ -1,5 +1,5 @@
 """Meet event parser classes and exceptions."""
 
-from .authentification import *
+from .authentication import *
 from .exceptions import *
 from .parsers import *
