+++ src/backend/core/recording/event/authentication.py
